+++ packages/client/components/bottom-drawer/index.tsx
@@ -2,9 +2,6 @@ import React, { useState } from 'react'
 import { Grid } from '@material-ui/core'
 import styled from 'styled-components'
 import { useSwipeable } from 'react-swipeable'
-import { Button } from '@material-ui/core'
-import { TextQuestionAnswer } from '../text-question-answer'
-import { Rating } from '../rating'
 import units from 'design-units'
 
 const Container = styled.div<{ drawerHeight: string }>`
@@ -48,7 +45,7 @@ const ChildContent = styled.div`
   height: 344px;
 `
 
-export interface BottomDrawerProps {
+export interface MobileDrawerProps {
   title: string
   children: React.ReactNode
   actionButton?: React.ReactNode
@@ -59,13 +56,13 @@ export interface BottomDrawerProps {
 // this components has 2 modes: adjustable height and fixed height
 // e.g. passing fixedHeight: '200px' will hide the button to adjust
 // hegith, so the drawer's height always stay the same
-export const BottomDrawer = ({
+export const MobileDrawer = ({
   title,
   children,
   actionButton,
   isClosed = false,
   fixedHeight
-}: BottomDrawerProps) => {
+}: MobileDrawerProps) => {
   const minimizedHeight = '52px'
   const expandedHeight = '400px'
   const [drawerHeight, setDrawerHeight] = useState(
@@ -100,55 +97,5 @@ export const BottomDrawer = ({
         <ChildContent>{children}</ChildContent>
       </Grid>
     </Container>
-  )
-}
-
-export interface FeedbackDrawerProps {
-  editable?: boolean
-  onSaveClick?: () => void
-  questions: string[]
-  answers: string[]
-}
-
-export const FeedbackDrawer = ({
-  onSaveClick = () => null,
-  questions,
-  answers,
-  editable
-}: FeedbackDrawerProps) => {
-  const [newAnswers, setNewAnswers] = useState([...answers])
-  const [isDirty, setIsDirty] = useState(false)
-
-  const saveButton = (
-    <Button
-      onClick={onSaveClick}
-      disabled={!isDirty}
-      color={isDirty ? 'primary' : 'default'}>
-      Save
-    </Button>
-  )
-
-  const setAnswer = (answer: string, i: number) => {
-    newAnswers[i] = answer
-    setNewAnswers(newAnswers)
-    setIsDirty(true)
-  }
-
-  const children = newAnswers.map((answer, i) => (
-    <TextQuestionAnswer
-      key={i}
-      question={questions[i]}
-      answer={answer}
-      editable={editable}
-      index={i}
-      onChange={setAnswer}
-    />
-  ))
-
-  return (
-    <BottomDrawer title={'Questions'} actionButton={saveButton}>
-      {children}
-      <Rating />
-    </BottomDrawer>
   )
 }
