+++ packages/client/components/feedback/stories.tsx
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/react'
-import { MobileDrawer } from '.'
+import { FeedbackDrawer } from '.'
 import { withKnobs, boolean, text } from '@storybook/addon-knobs'
 import styled from 'styled-components'
 import { Button } from '@material-ui/core'
@@ -20,11 +20,17 @@ const StoryContainer = styled.div`
 stories.add('default', () => {
   return (
     <StoryContainer>
-      <MobileDrawer
-        title={text('title', 'Drawer Title')}
-        actionButton={<Button color="primary">Button</Button>}>
-        some drawer content
-      </MobileDrawer>
+      <FeedbackDrawer
+        editable={boolean('editable', true)}
+        questions={[
+          'Did the author use meaningful and descriptive variable names?',
+          'Is the algorithm in function xxx efficient?'
+        ]}
+        answers={[
+          'Mostly yes, except for a few places the author used “xxx”. I think “yyy” would be more clear.',
+          'I think so. An alternative (equally efficient) way would be to xxx.'
+        ]}
+      />
     </StoryContainer>
   )
 })
